requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
newspaper3k>=0.2.8
requests>=2.31
python-dotenv>=1.0
Pillow>=10.0
feedparser>=6.0
lxml>=4.9
lxml-html-clean>=0.1
